Create MailApplication
A mail application using Python is a software program that allows users to create, send, and receive emails. It is typically built using the Python programming language and the smtplib module. The application typically has a user interface that allows users to compose emails, enter recipient addresses, and attach files. It also has a system for sending and receiving emails over the internet.